Add media query for tablet devices
I have added media query for tablet devices. So that, these devices can
be fully compatible on tablet screens.

Also I have fixed the minor bug, for my logo on the smaller resolutions
for my mobile devices so that it get's pushed on his own line.
I managed to fix that setting flex-basis selector to 100%.